- Use IndexRoute for react-router - move authenticateComponent to helpers - Update react-router, redux, redux-router
