+++ app/routes.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Route from 'react-router';
+import { Route, IndexRoute } from 'react-router';
 
 import App from 'containers/App';
 import Vote from 'containers/Vote';
@@ -7,11 +7,18 @@ import About from 'containers/About';
 import Login from 'containers/Login';
 import Dashboard from 'containers/Dashboard';
 
-import { requireAuthentication } from 'components/authenticateComponent';
+import { requireAuthentication } from 'helpers/authHelpers';
+
+function requireAuth(nextState, replace) {
+  replace({
+    pathname: '/login',
+    state: { nextPathname: nextState.location.pathname }
+  });
+}
 
 export default (
-  <Route component={App}>
-    <Route path="/" component={Vote} />
+  <Route path="/" component={App}>
+    <IndexRoute component={Vote} />
     <Route path="login" component={Login} />
     <Route path="dashboard" component={requireAuthentication(Dashboard)} />
     <Route path="about" component={About} />
