+++ package.json
@@ -71,12 +71,12 @@
     "react": "^0.14.2",
     "react-dom": "^0.14.2",
     "react-helmet": "2.1.1",
-    "react-redux": "^3.1.0",
-    "react-router": "^1.0.0",
+    "react-redux": "^4.0.6",
+    "react-router": "^1.0.3",
     "react-transform-catch-errors": "^1.0.0",
     "react-transform-hmr": "^1.0.1",
     "redbox-react": "^1.2.0",
-    "redux": "^3.0.2",
+    "redux": "^3.0.5",
     "redux-devtools": "^2.1.5",
     "redux-logger": "^2.0.3",
     "redux-simple-router": "^1.0.2",
